attributes & upgrades components

+++ src/main/java/com/jujutsu/systems/ability/upgrade/UpgradesData.java
@@ -9,17 +9,17 @@ import net.minecraft.util.Identifier;
 
 import java.util.HashMap;
 
-public record UpgradesData(Identifier upgradesId, float points, HashMap<Identifier, Identifier> purchasedUpgrades) {
+public record UpgradesData(Identifier upgradesId, int points, HashMap<Identifier, Identifier> purchasedUpgrades) {
     public static final Codec<UpgradesData> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             Identifier.CODEC.fieldOf("upgradesId").forGetter(UpgradesData::upgradesId),
-            Codec.FLOAT.fieldOf("points").forGetter(UpgradesData::points),
+            Codec.INT.fieldOf("points").forGetter(UpgradesData::points),
             Codec.unboundedMap(Identifier.CODEC, Identifier.CODEC).xmap(HashMap::new, HashMap::new)
                     .fieldOf("purchasedUpgrades").forGetter(UpgradesData::purchasedUpgrades)
     ).apply(instance, UpgradesData::new));
 
     public static final PacketCodec<RegistryByteBuf, UpgradesData> PACKET_CODEC = PacketCodec.tuple(
             Identifier.PACKET_CODEC, UpgradesData::upgradesId,
-            PacketCodecs.FLOAT, UpgradesData::points,
+            PacketCodecs.INTEGER, UpgradesData::points,
             PacketCodecs.map(HashMap::new, Identifier.PACKET_CODEC, Identifier.PACKET_CODEC), UpgradesData::purchasedUpgrades,
             UpgradesData::new);
 }
